+++ kernel/c/timeranger2/src/tr_treedb.h
@@ -171,6 +171,7 @@ PUBLIC int treedb_set_callback(
         "template"
         "uuid"
         "rowid"
+        "qualified"
         "password"
         "email"
         "url"